+++ packages/orq-rc/src/orq_ai_sdk/models/post_v2_feedback_evaluationop.py
@@ -55,7 +55,7 @@ class RequestBody3(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodyRequestSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-03-27T12:10:57.429Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-03-28T09:40:36.204Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -121,7 +121,7 @@ class RequestBody2(BaseModel):
 
     source: Optional[PostV2FeedbackEvaluationRequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-03-27T12:10:57.428Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-03-28T09:40:36.203Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -187,7 +187,7 @@ class RequestBody1(BaseModel):
 
     source: Optional[RequestBodySource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-03-27T12:10:57.428Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-03-28T09:40:36.202Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")