+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMQK8T9Y6743MWVERCRHYPJE"
+        "01KMSX27G0TVFR8P9XGXHRNC1C"
     )
     r"""The unique identifier of the data source"""